test: move global test file from test folder.
